pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dateno-cmd"
version = "0.1.0"
description = "Dateno command-line interface"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

authors = [
  { name = "Dateno Team", email = "dateno@dateno.io" }
]

dependencies = [
  "typer>=0.12",
  "PyYAML>=6.0",
  "tabulate>=0.9",
  "flatdict>=4.0",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",

  # SDK from git repository
  "dateno @ git+https://github.com/datenoio/dateno-sdk-python.git@main",
]

[project.scripts]
dateno = "dateno_cmd.cli:app"

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "pytest-cov>=4.0",
]

[tool.setuptools]
packages = [
  "dateno_cmd",
  "dateno_cmd.commands",
  "dateno_cmd.services",
  "dateno_cmd.utils",
]
